Removi as classes de CPF, CNPJ, ISBN10 e ISBN13, deu muito B.O. por causa dessas malditas classes separadas (vtnc claion)
