+++ .gitignore
@@ -1,1 +1,2 @@
 /utils/private
+tests
