+++ tests/teste.php
@@ -1,27 +1,3 @@
-<?php
-
-// // ini_set('display_errors', 1);
-// // ini_set('display_startup_errors', 1);
-// // error_reporting(E_ALL);
-
-// // require_once "../classe/DB.php";
-
-// // use classe\DB;
-
-// // $banco = DB::getInstance();
-
-// require_once "../utils/Email.php";
-
-// use classe\Email;
-
-// $obj = new Email();
-
-// // public function emailGenerico(string $assunto, string $conteudo, string $destino, string $alt){
-
-// //$obj->emailGenerico();
-
-?>
-
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -30,39 +6,40 @@
     <title>Document</title>
 </head>
 <body>
-    <!-- <form action="teste.php" method = "POST">
+    <form action="teste.php" method = "POST">
 
         Assunto: <input type="text" name = "assunto"><br>
         Conteúdo: <input type="text" name = "conteudo"><br>
         Destino: <input type="text" name = "destino"><br>
-        <!-- Assunto: <input type="text" name = "assunto"><br> -->
-        <!-- <input type="submit" name = "enviar" value = "Enviar"> -->
+        Assunto: <input type="text" name = "assunto"><br>
+        <input type="submit" name = "enviar" value = "Enviar">
 
-    <!-- </form>  -->
+    </form> 
 
     <?php
-        // if(isset($_POST["enviar"])){
-        //     if(isset($_POST["assunto"]) && !empty($_POST["assunto"])){
-        //         if(isset($_POST["conteudo"]) && !empty($_POST["conteudo"])){
-        //             if(isset($_POST["destino"]) && !empty($_POST["destino"])){
+        if(isset($_POST["enviar"])){
+            if(isset($_POST["assunto"]) && !empty($_POST["assunto"])){
+                if(isset($_POST["conteudo"]) && !empty($_POST["conteudo"])){
+                    if(isset($_POST["destino"]) && !empty($_POST["destino"])){
 
-        //                 $assunto = $_POST["assunto"];
-        //                 $conteudo = $_POST["conteudo"];
-        //                 $destino = $_POST["destino"];
+                        $assunto = $_POST["assunto"];
+                        $conteudo = $_POST["conteudo"];
+                        $destino = $_POST["destino"];
 
-        //                 $teste = $obj->emailGenerico($assunto, $conteudo, $destino);
+                        $teste = $obj->emailGenerico($assunto, $conteudo, $destino);
 
-        //                 if($teste == false){
+                        if($teste == false){
 
-        //                     echo "Email FALHOU!!<br>";
-        //                 }else{
+                                echo "Email FALHOU!!<br>";
 
-        //                     echo "Email foi enviado!";
-        //                 }
-        //             }
-        //         }
-        //     }
-        // }
+                            }else{
+                                
+                               echo "Email foi enviado!";
+                        }
+                    }
+                }
+            }
+        }
     
     ?> 
 
@@ -75,10 +52,11 @@
     ini_set('display_startup_errors', 1);
     error_reporting(E_ALL);
 
-    require_once "../utils/CPF.php";
-    use utils\CPF;
+    require_once "../utils/CodigoValido.php";
 
-    $obj = new CPF();
+    use classe\CodigoValido;
+
+    $obj = new CodigoValido();
 
     $teste = $obj->validarCPF("050.104.442.67");
